requirements.txt
Augmentor
kwargs
DisCapTy
topggpy==1.4.0
motor==2.5.1
algoliasearch==2.6.1
spotipy==2.19.0
dnspython==2.2.0
dislash.py==1.4.9
jishaku==2.3.2
vacefron.py==1.6.3
prawcore==2.3.0
pyjokes==0.6.0
googletrans==3.1.0a0
google==3.0.0
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
captcha
certifi==2021.5.30
dateparser==1.0.0
discord==1.7.3
discord.py==1.7.3
discord-components==2.0.5
discord-py-interactions==3.0.2
docutils==0.16
emoji==1.4.1
ffmpeg==1.4
multidict==5.1.0
lyrics-extractor==3.0.1
protobuf==3.17.3
discord-pretty-help==1.3.3
psutil==5.8.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
urllib3==1.26.6
wikipedia==1.4.0
yarl==1.6.3
youtube-dl==2021.3.3
IMDbPY==2021.4.18
nest_asyncio
tinydb
pyfiglet